Fix wrong colour for fullspaces
Another bug-fix: When the model was a homogenous fullspace, the xy- and
xz-plots had a wrong colour. yz was initially fine, but as soon as
scrolled would also have a wrong colour. The issue is with colorbar()
and only one value. So I introduced a +/- 1% range for the cases of
homogeneous fullspaces.