Improve load time
matplotlib, vtk, and omf are soft dependencies. However, if they are
installed they were, currently, loaded whether or not the mixins are
used. This adds significantly to the load overhead (>1s).

This commit makes them loading lazily, only if they are actually used.

It also adds a test, ensuring that the load-time of discretize is below
a threshold.