Allow __getitem__ to get int indices
Use `numpy.unravel_index` to convert integer indices into tuples,
following the right FORTRAN ordering. Simplify `__iter__` by building
the iterator with a simple for loop over the whole range of cells, and
reply on `__getitem__` for the unravelling of the indices. Add a new
test that checks if the int indices return the expected cells.